requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
google-generativeai>=0.7.2
